fix: 카카오 oauth2 인증시 id 값이 Int, Long 로 나뉘는 문제 수정

+++ src/main/kotlin/day/toons/global/config/security/oauth2/user/KakaoOAuth2UserInfo.kt
@@ -4,7 +4,7 @@ import day.toons.global.config.security.oauth2.user.OAuth2UserInfo
 
 class KakaoOAuth2UserInfo(attributes: MutableMap<String, Any>): OAuth2UserInfo(attributes) {
     override val id: String
-        get() = (attributes["id"] as Int).toString()
+        get() = (attributes["id"]).toString()
     override val name: String
         get() = (attributes["properties"] as Map<*, *>)["nickname"] as String
     override val email: String
